1. refactor dataflow_service.go 2. add heartbeat for message processing with nats
