+++ internal/common/task.go
@@ -30,4 +30,9 @@ type TaskHandle interface {
 	GetMessage() TaskMessage
 	Ack() error
 	Nack() error
+
+	// InProgress resets the AckWait timer without acknowledging the message,
+	// signalling the broker that the worker is still processing. Call
+	// periodically during long tasks to avoid in-flight redelivery.
+	InProgress() error
 }
