+++ internal/ingestion/task/chunk_process_test.go
@@ -201,12 +201,12 @@ func TestRenameTextToContentWithWeight_NoTextKey(t *testing.T) {
 }
 
 // =============================================================================
-// ProcessChunksForDataflow — Python: processChunks()
+// ProcessChunksForPipeline — Python: processChunks()
 // =============================================================================
 
-func TestProcessChunksForDataflow_SetsDocIDAndKBID(t *testing.T) {
+func TestProcessChunksForPipeline_SetsDocIDAndKBID(t *testing.T) {
 	chunks := []map[string]any{{"text": "hello world"}}
-	_ = ProcessChunksForDataflow(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
+	_ = ProcessChunksForPipeline(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
 
 	if chunks[0]["doc_id"] != "doc-1" {
 		t.Errorf("doc_id = %q, want \"doc-1\"", chunks[0]["doc_id"])
@@ -220,18 +220,18 @@ func TestProcessChunksForDataflow_SetsDocIDAndKBID(t *testing.T) {
 	}
 }
 
-func TestProcessChunksForDataflow_SetsDocNameKwd(t *testing.T) {
+func TestProcessChunksForPipeline_SetsDocNameKwd(t *testing.T) {
 	chunks := []map[string]any{{"text": "hello"}}
-	_ = ProcessChunksForDataflow(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
+	_ = ProcessChunksForPipeline(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
 	if chunks[0]["docnm_kwd"] != "test-doc.pdf" {
 		t.Errorf("docnm_kwd = %q, want \"test-doc.pdf\"", chunks[0]["docnm_kwd"])
 	}
 }
 
-func TestProcessChunksForDataflow_SetsTimeFields(t *testing.T) {
+func TestProcessChunksForPipeline_SetsTimeFields(t *testing.T) {
 	now := time.Now()
 	chunks := []map[string]any{{"text": "hello"}}
-	_ = ProcessChunksForDataflow(chunks, "doc-1", "kb-1", "test-doc.pdf", now)
+	_ = ProcessChunksForPipeline(chunks, "doc-1", "kb-1", "test-doc.pdf", now)
 
 	if timeStr, ok := chunks[0]["create_time"].(string); ok {
 		if timeStr != now.Format("2006-01-02 15:04:05") {
@@ -250,31 +250,31 @@ func TestProcessChunksForDataflow_SetsTimeFields(t *testing.T) {
 	}
 }
 
-func TestProcessChunksForDataflow_GeneratesID(t *testing.T) {
+func TestProcessChunksForPipeline_GeneratesID(t *testing.T) {
 	chunks := []map[string]any{{"text": "hello"}}
-	_ = ProcessChunksForDataflow(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
+	_ = ProcessChunksForPipeline(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
 	id, ok := chunks[0]["id"].(string)
 	if !ok || id == "" {
 		t.Errorf("id should be non-empty string, got %v", chunks[0]["id"])
 	}
 }
 
-func TestProcessChunksForDataflow_NoPanicOnListText(t *testing.T) {
+func TestProcessChunksForPipeline_NoPanicOnListText(t *testing.T) {
 	chunks := []map[string]any{{"text": []any{"bad-shape"}}}
-	res := ProcessChunksForDataflow(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
+	res := ProcessChunksForPipeline(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
 	if res == nil {
 		t.Errorf("should return valid result")
 	}
 }
 
-func TestProcessChunksForDataflow_RemovesInternalPipelineFields(t *testing.T) {
+func TestProcessChunksForPipeline_RemovesInternalPipelineFields(t *testing.T) {
 	chunks := []map[string]any{{
 		"text":           "hello",
 		"_pdf_positions": []any{[]any{0, 1, 2, 3, 4}},
 		"image":          "data:image/png;base64,abc",
 	}}
 
-	_ = ProcessChunksForDataflow(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
+	_ = ProcessChunksForPipeline(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
 	if _, exists := chunks[0]["_pdf_positions"]; exists {
 		t.Fatalf("_pdf_positions should be removed before indexing: %v", chunks[0]["_pdf_positions"])
 	}
@@ -283,17 +283,17 @@ func TestProcessChunksForDataflow_RemovesInternalPipelineFields(t *testing.T) {
 	}
 }
 
-func TestProcessChunksForDataflow_PreservesExistingID(t *testing.T) {
+func TestProcessChunksForPipeline_PreservesExistingID(t *testing.T) {
 	chunks := []map[string]any{{"text": "hello", "id": "existing-id"}}
-	_ = ProcessChunksForDataflow(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
+	_ = ProcessChunksForPipeline(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
 	if chunks[0]["id"] != "existing-id" {
 		t.Errorf("existing id should be preserved, got %q", chunks[0]["id"])
 	}
 }
 
-func TestProcessChunksForDataflow_QuestionsProcessing(t *testing.T) {
+func TestProcessChunksForPipeline_QuestionsProcessing(t *testing.T) {
 	chunks := []map[string]any{{"text": "hello", "questions": "Q1\nQ2\nQ3"}}
-	_ = ProcessChunksForDataflow(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
+	_ = ProcessChunksForPipeline(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
 
 	if _, exists := chunks[0]["questions"]; exists {
 		t.Error("questions key should be removed")
@@ -310,9 +310,9 @@ func TestProcessChunksForDataflow_QuestionsProcessing(t *testing.T) {
 	}
 }
 
-func TestProcessChunksForDataflow_KeywordsProcessing(t *testing.T) {
+func TestProcessChunksForPipeline_KeywordsProcessing(t *testing.T) {
 	chunks := []map[string]any{{"text": "hello", "keywords": "kw1,kw2;kw3"}}
-	_ = ProcessChunksForDataflow(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
+	_ = ProcessChunksForPipeline(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
 
 	if _, exists := chunks[0]["keywords"]; exists {
 		t.Error("keywords key should be removed")
@@ -326,9 +326,9 @@ func TestProcessChunksForDataflow_KeywordsProcessing(t *testing.T) {
 	}
 }
 
-func TestProcessChunksForDataflow_SummaryProcessing(t *testing.T) {
+func TestProcessChunksForPipeline_SummaryProcessing(t *testing.T) {
 	chunks := []map[string]any{{"text": "hello", "summary": "This is a summary."}}
-	_ = ProcessChunksForDataflow(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
+	_ = ProcessChunksForPipeline(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
 
 	if _, exists := chunks[0]["summary"]; exists {
 		t.Error("summary key should be removed")
@@ -341,9 +341,9 @@ func TestProcessChunksForDataflow_SummaryProcessing(t *testing.T) {
 	}
 }
 
-func TestProcessChunksForDataflow_TextRenamed(t *testing.T) {
+func TestProcessChunksForPipeline_TextRenamed(t *testing.T) {
 	chunks := []map[string]any{{"text": "hello world"}}
-	_ = ProcessChunksForDataflow(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
+	_ = ProcessChunksForPipeline(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
 
 	if _, exists := chunks[0]["text"]; exists {
 		t.Error("text key should be removed")
@@ -353,9 +353,9 @@ func TestProcessChunksForDataflow_TextRenamed(t *testing.T) {
 	}
 }
 
-func TestProcessChunksForDataflow_PreservesContentWithWeight(t *testing.T) {
+func TestProcessChunksForPipeline_PreservesContentWithWeight(t *testing.T) {
 	chunks := []map[string]any{{"content_with_weight": "already set", "text": "hello"}}
-	_ = ProcessChunksForDataflow(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
+	_ = ProcessChunksForPipeline(chunks, "doc-1", "kb-1", "test-doc.pdf", time.Now())
 	if chunks[0]["content_with_weight"] != "already set" {
 		t.Errorf("content_with_weight = %q, want \"already set\"", chunks[0]["content_with_weight"])
 	}
