+++ internal/ingestion/task/task_context.go
@@ -21,6 +21,7 @@ import (
 	"fmt"
 	"strings"
 
+	"ragflow/internal/common"
 	"ragflow/internal/dao"
 	"ragflow/internal/entity"
 )
@@ -39,6 +40,12 @@ type TaskContext struct {
 	File       any
 
 	ProgressFunc ProgressFunc
+
+	// Handle is the message-queue ack handle for the task message that scheduled
+	// this context. The scheduler sets it before queueing; the worker acks on a
+	// durably-persisted terminal status and nacks otherwise (e.g. shutdown
+	// mid-task) so the message is redelivered and resumed after restart.
+	Handle common.TaskHandle
 }
 
 // NewTaskContextForScheduling creates a lightweight TaskContext for queue scheduling.
